refactor(assay): create internal/ directory, extract util.rs
