+++ src/commands/assay/internal/mod.rs
@@ -1,0 +1,7 @@
+//! Internal implementation for assay command
+//!
+//! Follows dependable-rust pattern: private modules with curated re-exports.
+
+mod util;
+
+pub(super) use util::truncate;
